feat(voting-event-apis): apis called as rest service are always checked for authorisation
apis called from service.ts are always checked for the authorization of the user calling them. A
different set of functions (e.g. openVotingEventVerified) is added for the cases when these apis
have to be called from operations internal to the server. For intance, the cancelInitiative api is
always checked for authorization, but once the operation is authorized then the operations to cancel
all voting events belonging to the initiative to be cancelled are to be considered authorized and
therefore the cancelVotingEventVerified function is called.