feat(executeMongoService): hack to overcome the issue with authorization header and lambda
Currently I am not able to pass the authorization header to the backend when the back end runs as a
lambda function. The header contains the token which, when decripted, provides to the backend the
user which is currently logged. Since such authorization token does not currently reach the lambda
function , as a hack, I pass the user as part of the payload. This has to be removed as soon as we
find a way to pass the authorization token to the lambda function.